feat(#87): refactor: lsp-auditor.ts — audit and modularize
