+++ test/lsp-auditor-settings.test.mts
@@ -1,11 +1,8 @@
 /**
  * Phase 3: Settings/config loading tests for LSP Auditor
  *
- * Tests buildServerMappings from lsp-auditor.ts with various
- * settings.json configurations.
- *
- * Functions duplicated from lsp-auditor.ts (pattern from test/session-logger.test.mts,
- * test/supervisor-extensions.test.mts).
+ * Tests buildServerMappings from .pi/extensions/lsp-auditor/server-mappings.ts
+ * with various settings.json configurations.
  *
  * Run with:
  *   node --experimental-strip-types --test test/lsp-auditor-settings.test.mts
@@ -13,64 +10,7 @@
 
 import assert from "node:assert";
 import { describe, it } from "node:test";
-
-// ═══════════════════════════════════════════════════════════════════════
-// Duplicated from lsp-auditor.ts for pure unit testing
-// ═══════════════════════════════════════════════════════════════════════
-
-interface ServerMapping {
-	extensions: string[];
-	command: string;
-	args: string[];
-	severityThreshold: "error" | "warning" | "info";
-}
-
-const DEFAULT_SERVER_MAPPINGS: ServerMapping[] = [
-	{ extensions: [".ts", ".tsx", ".js", ".jsx"], command: "typescript-language-server", args: ["--stdio"], severityThreshold: "warning" },
-	{ extensions: [".py"], command: "pyright-langserver", args: ["--stdio"], severityThreshold: "warning" },
-	{ extensions: [".rs"], command: "rust-analyzer", args: [], severityThreshold: "warning" },
-	{ extensions: [".go"], command: "gopls", args: [], severityThreshold: "warning" },
-];
-
-function buildServerMappings(configRaw: unknown): ServerMapping[] {
-	if (!configRaw || typeof configRaw !== "object") return [...DEFAULT_SERVER_MAPPINGS];
-
-	const config = configRaw as { servers?: Array<{ extensions: string[]; command: string; args?: string[]; severityThreshold?: string }> };
-	if (!config.servers || !Array.isArray(config.servers) || config.servers.length === 0) return [...DEFAULT_SERVER_MAPPINGS];
-
-	const merged = [...DEFAULT_SERVER_MAPPINGS];
-
-	for (const srv of config.servers) {
-		if (!srv.extensions || !Array.isArray(srv.extensions) || srv.extensions.length === 0) continue;
-		if (!srv.command || typeof srv.command !== "string" || !srv.command.trim()) continue;
-
-		const exts = [...new Set(srv.extensions.map(e => e.toLowerCase()))];
-
-		let threshold: "error" | "warning" | "info" = "warning";
-		if (srv.severityThreshold) {
-			const t = srv.severityThreshold.toLowerCase();
-			if (t === "error" || t === "warning" || t === "info") threshold = t;
-		}
-
-		const newMapping: ServerMapping = {
-			extensions: exts,
-			command: srv.command.trim(),
-			args: srv.args || [],
-			severityThreshold: threshold,
-		};
-
-		const overlapExts = new Set(exts);
-		for (let i = merged.length - 1; i >= 0; i--) {
-			if (merged[i]!.extensions.some(e => overlapExts.has(e.toLowerCase()))) {
-				merged.splice(i, 1);
-			}
-		}
-
-		merged.push(newMapping);
-	}
-
-	return merged;
-}
+import { buildServerMappings } from "../.pi/extensions/lsp-auditor/server-mappings.ts";
 
 // ═══════════════════════════════════════════════════════════════════════
 // Tests
